requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
channels>=4.0.0
channels_redis>=4.1.0
psycopg2-binary>=2.9.9
daphne>=4.0.0
redis>=5.0.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
whitenoise>=6.5.0
